requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-ollama==0.2.3
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.7
openai==1.63.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
